massively improved typing
