+++ src/typing/context.rs
@@ -1,11 +1,35 @@
-use std::collections::hash_map::HashMap;
+use std::collections::{HashMap, HashSet};
 use crate::ast::common::Ident;
 use crate::ast::typed_rust::PostType;
 
 #[derive(Clone, Debug)]
 pub struct GlobalContext {
-    enums : HashMap<Ident, Vec<Ident>>,
-    structs : HashMap<Ident, Vec<(Ident, PostType)>>,
+    enums : HashMap<String, HashSet<String>>,
+    structs : HashMap<String, HashMap<String, PostType>>,
+    known_types : HashMap<String, PostType>,
+}
+
+impl GlobalContext {
+
+    pub fn new() -> Self {
+        Self {
+            enums : HashMap::new(),
+            structs : HashMap::new(),
+            known_types : HashMap::new(),
+        }
+    }
+
+    pub fn insert(&mut self, name : String, typ : PostType) -> Option<PostType> {
+        self.known_types.insert(name, typ)
+    }
+
+    pub fn get_known_types(&self) -> &HashMap<String, PostType> {
+        &self.known_types
+    }
+
+    pub fn get_typ(&self, name : &str) -> Option<&PostType> {
+        self.known_types.get(name)
+    }
 }
 
 
@@ -16,6 +40,13 @@ pub struct LocalContext {
 }
 
 impl LocalContext {
+
+    pub fn new() -> Self {
+        Self {
+            vars : Vec::new()
+        }
+    }
+
     pub fn add_layer(&mut self) {
         self.vars.push(HashMap::new())
     }
